+++ e2e/examples/nextjs-node-runtime/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @llamaindex/next-node-runtime
 
+## 0.1.16
+
+### Patch Changes
+
+- 4999df1: bump nextjs
+- Updated dependencies [f5e4d09]
+  - llamaindex@0.9.16
+
 ## 0.1.15
 
 ### Patch Changes