+++ e2e/examples/waku-query-engine/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @llamaindex/waku-query-engine-test
 
+## 0.0.150
+
+### Patch Changes
+
+- Updated dependencies [f5e4d09]
+  - llamaindex@0.9.16
+
 ## 0.0.149
 
 ### Patch Changes